feat: 커스텀 훅 useSearchMovies 구현 및 검색 결과 로직 분리
